requirements.txt
pandas
click
django
djangorestframework
